requirements/requirements.txt
django==3.2.16
django-environ==0.9.0
mysqlclient==1.4.6
djangorestframework==3.14.0
